vault/requirements.txt
cffi==1.15.0
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf==3.20.1
pycparser==2.21
PyNaCl==1.5.0
six==1.16.0
